Lifestyle_App/app/src/androidTest/java/com/example/willdunn/lifestyle_app/LoginFragmentTest.java
package com.example.willdunn.lifestyle_app;

import android.view.View;

import com.example.willdunn.lifestyle_app.View.LaunchActivity;
import com.example.willdunn.lifestyle_app.View.LoginFragment;

import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;

import androidx.test.rule.ActivityTestRule;

import static androidx.test.espresso.Espresso.onView;
import static androidx.test.espresso.action.ViewActions.click;
import static androidx.test.espresso.action.ViewActions.replaceText;
import static androidx.test.espresso.action.ViewActions.typeText;
import static androidx.test.espresso.assertion.ViewAssertions.matches;
import static androidx.test.espresso.matcher.ViewMatchers.withClassName;
import static androidx.test.espresso.matcher.ViewMatchers.withId;
import static androidx.test.espresso.matcher.ViewMatchers.withText;
import static org.hamcrest.Matchers.is;
import static org.hamcrest.core.AllOf.allOf;
import static org.hamcrest.core.StringEndsWith.endsWith;
import static org.junit.Assert.assertNotNull;

public class LoginFragmentTest {

    @Rule
    public ActivityTestRule<LaunchActivity> launchActivityActivityTestRule = new ActivityTestRule<>(LaunchActivity.class);

    private LaunchActivity launchActivity = null;

    private LoginFragment login_frag = null;

    private View v = null;

    @Before
    public void setUp() throws Exception {
        launchActivity = launchActivityActivityTestRule.getActivity();
        login_frag = new LoginFragment();
        launchActivity.getSupportFragmentManager().beginTransaction().add(login_frag.getId(), login_frag).commit();

    }

    @Test
    public void testLaunch() {
        assertNotNull(login_frag);
    }

    // The below tests (the commented out ones) don't work any longer because we have designated certain
    // threads to run on non-UI threads, but the tests don't reflect that and will fail due to threading issues.
//    @Test
//    public void onClickSignUp() {
//        launchActivity.runOnUiThread(new Runnable() {
//            public void run() {
//                View vSign = login_frag.getView().findViewById(R.id.sign_up);
//                login_frag.onClick(vSign);
//            }
//        });
//
//        onView(withId(R.id.frag_signup)).check(matches((isDisplayed())));
//    }
//
//    @Test
//    public void onClickLoginSuccess() {
//        launchActivity.runOnUiThread(new Runnable() {
//            public void run() {
//                EditText userName = login_frag.getView().findViewById(R.id.user_name);
//                EditText password = login_frag.getView().findViewById(R.id.password);
//                userName.setText("default");
//                password.setText("password");
//                View vLog = login_frag.getView().findViewById(R.id.login);
//                login_frag.onClick(vLog);
//            }
//        });
//
//        onView(withId(R.id.frag_regimen)).check(matches((isDisplayed())));
//    }
//
//    @Test
//    public void onClickLoginFailed() {
//
//        launchActivity.runOnUiThread(new Runnable() {
//            public void run() {
//                EditText userName = login_frag.getView().findViewById(R.id.user_name);
//                EditText password = login_frag.getView().findViewById(R.id.password);
//                userName.setText("dev");
//                password.setText("badpassword");
//                View vLog = login_frag.getView().findViewById(R.id.login);
//                login_frag.onClick(vLog);
//            }
//        });
//
//        onView(withText("Username/Password Incorrect")).inRoot(withDecorView(not(is(launchActivityActivityTestRule.getActivity().getWindow().getDecorView())))).check(matches(isDisplayed()));
//    }


    @Test
    public void validateEnteredUsername() {
        onView(allOf(withClassName(endsWith("EditText")), withText(is("will dunn"))))
                .perform(replaceText(""));
        onView(withId(R.id.user_name)).perform(typeText("default"), click());
        onView(withId(R.id.user_name)).check(matches(withText("default")));
    }

    @Test
    public void validateEnteredPassword() {
        onView(allOf(withClassName(endsWith("EditText")), withText(is("password"))))
                .perform(replaceText(""));
        onView(withId(R.id.password)).perform(typeText("this is only a test"), click());
        onView(withId(R.id.password)).check(matches(withText("this is only a test")));
    }

}